+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Unreleased
 
+- **[NEW]** Add the ItemAction component
+- **[NEW]** Add href and onClick on Profile
+- **[UPDATE]** Add the hover state on Toggle button
+- **[UPDATE]** Add the hover state on profile
+- **[BREAKING CHANGE]** Add horizontal padding on Toggle button
+- **[BREAKING CHANGE]** Add horizontal padding on profile
 - [...]
 
 # v0.20.7 (18/01/2019)